Moved sound control functions to trackControl.py and added unit tests.
