+++ Logs/OutputLog.txt
@@ -1,1 +1,1 @@
-0:Beat Of Success.mp3,1:bensound-dubstep.wav,2:The Swag Is Back.mp3,Now playing:Music/The Swag Is Back.mp3,0:Beat Of Success.mp3,1:bensound-dubstep.wav,2:The Swag Is Back.mp3,Now playing:Music/bensound-dubstep.wav,Song stopped.,
+0:Beat Of Success.mp3,1:bensound-dubstep.wav,2:The Swag Is Back.mp3,Now playing: Music/bensound-dubstep.wav,0:Beat Of Success.mp3,1:bensound-dubstep.wav,2:The Swag Is Back.mp3,Now playing: Music/The Swag Is Back.mp3,Song stopped.,0:Beat Of Success.mp3,1:bensound-dubstep.wav,2:The Swag Is Back.mp3,Now playing: Music/The Swag Is Back.mp3,0:Beat Of Success.mp3,1:bensound-dubstep.wav,2:The Swag Is Back.mp3,Now playing: Music/The Swag Is Back.mp3,Song stopped.,
